% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{bs}
\alias{bs}
\title{Black Scholes Calculator}
\source{
\url{http://en.wikipedia.org/wiki/Black-Scholes_model}
		
		Options, Futures and Other Derivatives (6th Edition) by John C. Hull
}
\usage{
bs(S, K, r, vol, mat)
}
\arguments{
\item{S}{Current stock price}

\item{K}{Strike price}

\item{r}{Risk free rate of return}

\item{vol}{Volatility of the stock expressed as \eqn{\sigma}{\sigma}}

\item{mat}{Time to maturity}
}
\description{
Uses black Scholes formula to calculate the value of an option.
}
\details{
You need to convert \code{r} and \code{vol} to match the time period
		\code{mat}.
}
\note{
Assumes that the option is european, there is a constant risk-free
     interest rate, the stock price follows geometric Brownian motion with
     constant drift and volatility, and the stock does not pay a dividend.
     Also assumes that there are no transaction costs and the market is
     frictionless.
}
\author{
Gene Leynes
}
\seealso{
\code{\link{bs_withdiv}},
}

