% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_withdiv.R
\name{bs_withdiv}
\alias{bs_withdiv}
\title{Black Scholes Calculator (with dividend)}
\source{
\url{http://en.wikipedia.org/wiki/Black-Scholes_model}

		Options, Futures and Other Derivatives (6th Edition) by John C. Hull
}
\usage{
bs_withdiv(S, K, r, vol, mat, q = 0)
}
\arguments{
\item{S}{Current stock price}

\item{K}{Strike price}

\item{r}{Risk free rate of return}

\item{vol}{Volatility of the stock expressed as \eqn{\sigma}{\sigma}}

\item{mat}{Time to maturity}

\item{q}{Dividend rate, continuous}
}
\description{
Uses black Scholes formula to calculate the value of an option.
}
\details{
You need to convert \code{r} and \code{vol} to match the time period
		\code{mat}.
}
\note{
Assumes that the option is european, there is a constant risk-free
     interest rate, the stock price follows geometric Brownian motion with
     constant drift and volatility, and the stock does not pay a dividend.
     Also assumes that there are no transaction costs and the market is
     frictionless.
}
\examples{
		bs_withdiv(S=370, K=485, r=.04, vol=.1, mat=2, q=0)
		bs_withdiv(S=370, K=485, r=.04, vol=.1, mat=2, q=.01)


}
\seealso{
\code{\link{bs}}
}
\author{
Gene Leynes
}
