% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Iterative algorithm to determine the center of a polygon
find the center of mass / furthest away from any boundary

Takes as input a spatial polygon}
\arguments{
\item{pol}{One or more polygons as input}

\item{ultimate}{optional Boolean, TRUE = find polygon furthest away from centroid. False = ordinary centroid

## Source for functions:
## https://gis.stackexchange.com/a/265475/78424}
}
\description{
Iterative algorithm to determine the center of a polygon
find the center of mass / furthest away from any boundary

Takes as input a spatial polygon
}
