% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chi_census_tracts_2000}
\alias{chi_census_tracts_2000}
\title{Chicago Census Tracts 2000 Census}
\format{
A SpatialPolygonsDataFrame with 5 slots and 878 Polygons:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
chi_census_tracts_2000
}
\description{
Chicago census tracts for 2000
}
\keyword{datasets}
