% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chi_census_tracts_2010}
\alias{chi_census_tracts_2010}
\title{chi_census_tracts_2010}
\format{
A SpatialPolygonsDataFrame with 5 slots and 801 Polygons:
\describe{
  \item{statefp10}{The state code for Illinois; 17}
  \item{countyfp10}{The county code for Cook; 31}
  \item{geoid10}{The 10 digit GEOID, e.g. 17031760803}
  \item{tractce}{The last 6 digits of geoid10, e.g. 17031760803}
  \item{name10}{Tract / block group, e.g. 7608.03}
  \item{namelsad10}{Tract / block group with "Census Tract", e.g. "Census Tract 7608.03"}
  \item{commarea_n}{Community area number (1-77)}
  \item{commarea}{identical to commarea_n}
  \item{notes}{Notes}
  ...
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Census-Tracts-2010/5jrd-6zik}
}
\usage{
chi_census_tracts_2010
}
\description{
Spatial data containing boundaries for Chicago's 2010 census tracts.
}
\keyword{datasets}
