% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chi_community_areas}
\alias{chi_community_areas}
\title{Chicago's 77 Community Areas}
\format{
A SpatialPolygonsDataFrame with 5 slots and 878 Polygons:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6}
}
\usage{
chi_community_areas
}
\description{
Chicago census tracts for 2000
}
\keyword{datasets}
