% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chi_wards_2015}
\alias{chi_wards_2015}
\title{chi_wards}
\format{
A SpatialPolygonsDataFrame with 5 slots and 878 Polygons:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Wards-2015-/sp34-6z76}
}
\usage{
chi_wards_2015
}
\description{
Spatial data set for Chicago's 50 wards, as of 2015
}
\keyword{datasets}
