% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chi_zip_codes}
\alias{chi_zip_codes}
\title{Chicago zip codes}
\format{
A SpatialPolygonsDataFrame with 5 slots and 61 Polygons:
\describe{
  \item{objectid}{?}
  \item{shap_area}{num: shape area}
  \item{shape_len}{num: shape boundary length}
  \item{zip}{char: zipcode}
  ...
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-ZIP-Codes/gdcf-axmw}
}
\usage{
chi_zip_codes
}
\description{
Spatial data for Chicago's zip codes
}
\keyword{datasets}
