% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipped.R
\name{clipped}
\alias{clipped}
\title{Dump clipboard contents into R; especially from Excel (Windows)}
\usage{
clipped(
  header = TRUE,
  sep = "\\t",
  na.strings = c("", "NA"),
  check.names = FALSE,
  stringsAsFactors = F,
  dec = ".",
  as.data.table = TRUE,
  remove.spaces = TRUE,
  ...
)
}
\arguments{
\item{lowercase}{Converts column names to lower case}

\item{removedot}{Removes .'s from column names}
}
\description{
Useful shortcut for reading objects from your clipboard.
}
\details{
Only works in Windows. Puts everything in a data.frame, uses
		\code{stringsAsFactors=FALSE}
}
\examples{
	\dontrun{

     require(geneorama)

	## Copy a range of cells in Excel (or whatever) then assign the data
	## to "clipped_example"
	clipped_example <- clipped()

	## View the result in R
	print(clipped_example)
	str(clipped_example)
	}

## New version based on:
## https://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r

}
\seealso{
\code{\link{clipboard}},
}
\author{
Gene Leynes
}
