% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_datatable_FactorString.R
\name{convert_datatable_FactorString}
\alias{convert_datatable_FactorString}
\title{Convert data.table factor columns to string columns}
\usage{
convert_datatable_FactorString(dat, cols = NULL)
}
\arguments{
\item{dat}{A data.table}

\item{cols}{(optional) A vector of column names or column numbers to be 
converted. If blank, then all columns of class character are
converted.}
}
\description{
Shortcut to convert data.table columns from factor to string.
}
\details{
Returns the data.table invisibly because data.tables are modified in 
     place, and there is no need to make a copy.
}
\examples{
		require(geneorama)
		## Create examples
		dt <- as.data.table(OrchardSprays)
		dt[ , rowpos := as.factor(rowpos)]
		dt[ , colpos := as.factor(colpos)]
		str(dt)
		dt2 <- copy(dt)
		
		## No columns specified
		convert_datatable_FactorString(dt)
		str(dt)
		## Specify column by position
		convert_datatable_FactorString(dt2, cols=2)
		str(dt2)
		convert_datatable_FactorString(dt2, cols="colpos")
		str(dt2)
	
}
\seealso{
\code{\link[data.table]{data.table}}
}
\author{
Gene Leynes
}
