% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_datatable_int_to_num.R
\name{convert_datatable_IntNum}
\alias{convert_datatable_IntNum}
\title{Find and convert data.table columns from integer to numeric}
\usage{
convert_datatable_IntNum(dat, cols = NULL)
}
\arguments{
\item{dat}{A data.table}

\item{cols}{(optional) A subset of columns to search (see details)}
}
\description{
Converts a data.table's columns from integer to numeric. This conversion
		is done BY REFERENCE meaning that the original data.table is modified.
}
\details{
The data.table is returned invisibly if needed.

		The cols can be specified as a character vector of column names.  The
		entire data.table is returned, but only the columns in "cols" are
		searched.  A warning is produced if no columns are integer.
}
\author{
Gene Leynes
}
\seealso{
\code{\link[data.table]{data.table}}
	\code{\link[data.table]{set}}
}

