% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_datatable_StringDate.R
\name{convert_datatable_StringDate}
\alias{convert_datatable_StringDate}
\title{Find and convert data.table columns from Strings to IDate}
\usage{
convert_datatable_StringDate(dat, cols, fmt = NULL, tz = "GMT",
  newclass = c("POSIXct", "IDate", "ITime")[2])
}
\arguments{
\item{dat}{A data.table}

\item{cols}{An OPTIONAL subset of columns to search (see details)}

\item{fmt}{Format for conversion, default assumes "YYYY-MM-DD HH:MM:SS"
aka "\%Y-\%m-\%d \%H:\%M:\%S"}

\item{tz}{Timezone to convert to (only used for POSIX).  Default is "GMT"}
}
\description{
Converts a data.table's columns from strings to to IDates.  This
		conversion is done BY REFERENCE meaning that the original data.table is
		modified.
}
\details{
The data.table is returned invisibly if needed.

		Only POSIXct is used.

		The cols can be specified as a character vector of column names.  The
		entire data.table is returned, but only the columns in "cols" are
		searched.  A warning is produced if no columns are character.
}
\author{
Gene Leynes
}
\seealso{
\code{\link[data.table]{data.table}}
	\code{\link[data.table]{as.IDate}}
	\code{\link[data.table]{set}}
	\code{\link[base]{strptime}}
}

