% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_datatable_StringFactor.R
\name{convert_datatable_StringFactor}
\alias{convert_datatable_StringFactor}
\title{Convert data.table character columns to factor columns}
\usage{
convert_datatable_StringFactor(dat, cols = NULL)
}
\arguments{
\item{dat}{A data.table}

\item{cols}{(optional) A vector of column names or column numbers to be 
converted. If blank, then all columns of class character are
converted.}
}
\description{
Shortcut to convert data.table columns from string to factor.
}
\details{
Returns the data.table invisibly because data.tables are modified in 
     place, and there is no need to make a copy.
}
\examples{
		require(geneorama)
		## Create examples
		dt <- as.data.table(OrchardSprays)
		dt[ , treatment := as.character(treatment)]
		dtchar <- copy(dt)
		dtchar[ , rowpos := as.character(rowpos)]
		dtchar[ , colpos := as.character(colpos)]
		str(dt)
		str(dtchar)
		
		## No columns specified
		convert_datatable_StringFactor(dt)
		str(dt)
		## Specify column by position
		convert_datatable_StringFactor(dtchar, cols=2)
		str(dtchar)
		convert_datatable_StringFactor(dtchar, cols="colpos")
		str(dtchar)
	
}
\author{
Gene Leynes
}
\seealso{
\code{\link[data.table]{data.table}}
}

