% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cook_dupage_tracts_2018}
\alias{cook_dupage_tracts_2018}
\title{cook_dupage_tracts_2018}
\format{
A SpatialPolygonsDataFrame with 5 slots and 878 Polygons:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
cook_dupage_tracts_2018
}
\description{
Chicago census tracts for 2000
}
\keyword{datasets}
