% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cook_dupage_tracts_2019}
\alias{cook_dupage_tracts_2019}
\title{chi_census_tracts_2000}
\format{
A SpatialPolygonsDataFrame with 5 slots and 878 Polygons:
\describe{
  \item{objectid}{Object id}
  \item{tract_cens}{Census year}
  \item{census_t_1}{10 digit GEOID}
  \item{tract_fips}{State / County (17031)}
  \item{census_tra}{Remaining 6 digits of geoid}
  \item{tract_numa}{}
  \item{tract_comm}{Community area number}
  \item{tract_cr_1}{blank}
  \item{data_admin}{blank}
  \item{tract_crea}{blank}
  \item{perimeter}{blank}
  \item{tract_ce_1}{Numeric - ?}
  \item{tract_cent}{Numeric - ?}
  \item{shape_area}{Numeric – Shape area}
  \item{shape_len}{Numeric – Shape length?}
  \item{tract_ce_3}{Numeric – Longitude}
  \item{tract_ce_2}{Numeric – Latitude}
  ...
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Census-Tracts-2000/pt6c-hxpp}
}
\usage{
cook_dupage_tracts_2019
}
\description{
Chicago census tracts for 2000
}
\keyword{datasets}
