% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencylist.R
\name{dependencylist}
\alias{dependencylist}
\title{List all the dependent packages of the currently attached packages}
\usage{
dependencylist(
  standard_packages = c("stats", "graphics", "grDevices", "utils", "datasets",
    "methods", "base"),
  exclude_standard = TRUE
)
}
\arguments{
\item{standard_packages}{Packages considered by the user to be "standard".
The defaults is  c('stats', 'graphics', 'grDevices', 
                   'utils', 'datasets', 'methods', 
                   'base')}

\item{exclude_standard}{Exclude standard_packages in list? Default is TRUE}
}
\description{
Get a list where each element is a currently attached pacakge.  The
		elements of that list are all the dependent pacakges.
}
\details{
This function could change in the future.
}
\examples{
		dependencylist()

}
\seealso{
\code{\link[base]{loadedNamespaces}}
		\code{\link[utils]{installed.packages}}
}
\author{
Gene Leynes
}
