% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dtconv}
\alias{dtconv}
\title{dtconv}
\usage{
dtconv(dt, newclass, columns, copy = FALSE, dateformat = NULL,
  dateorigin = NULL, remove_pattern = NULL)
}
\arguments{
\item{dt}{A data.table to be converted by reference (the default)}

\item{newclass}{The desired new class}

\item{columns}{The columns to be converted}

\item{copy}{Should dtconv make a copy of dt? Default value is FALSE,
                 i.e. the data.table is modified by REFERENCE}
}
\description{
Convert columns of data.table objects.

		This function easily converts the column classes of data.table objects by
		passing in the column names, and the desired new class.
}
\author{
Gene Leynes
}

