\name{geneorama-package}
\alias{geneorama-package}
\alias{geneorama}
\docType{package}
\title{
Gene's random collection of functions
}
\description{
Some things that I've accumulated over the years.
}
\details{
\tabular{ll}{
Package: \tab geneorama\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-07\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
