% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.R
\name{heatplot}
\alias{heatplot}
\title{Make colorful scatter plots where the color show relative values of a third dimension}
\usage{
heatplot(
  x,
  y,
  z,
  bgcol = "#777044",
  coltype = "heat",
  ccex = 1.5,
  circles = TRUE,
  ...
)
}
\arguments{
\item{x}{vector for x position}

\item{y}{vector for y position}

\item{z}{vector for point size}

\item{bgcol}{Background color.  Default is a brown color: "#777044"}

\item{coltype}{What family of colors to use.  Default is "heat"}

\item{ccex}{Controls point size.  Default is 1.5}

\item{circles}{Do you want circles drawn around the plots?  (can be bad
when there are a lot of points) Defaults to TRUE}

\item{...}{Additional parameters passed to plot()}
}
\description{
Simple way to make scatter plot heatmaps
}
\details{
Plots x-y points with color to indicate size.
}
\examples{
N <- 1e4
DT <- data.table(x <- 1:N,
				 y <- -exp(-x/3000-100) * 5e44 + 30 +
				 		 	rnorm(N, sd = 1 + (log(x/10)*.3)),
				 z <- sqrt(y) + rnorm(N, sd = 1 + (x*.03)))
DT[ ,plot(x,y)]
DT[ ,plot(x,z)]
DT[ ,heatplot(x,y,z)]
DT[ ,heatplot(x,y,z, circles=FALSE)]


}
\author{
Gene Leynes
}
