% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebm.R
\name{makebm}
\alias{makebm}
\title{Make a matrix simulating Brownian motion}
\usage{
makebm(mu = 0.05, sigma = 0.2, dt = 1, T = 10, sims = 10000)
}
\arguments{
\item{mu}{Drift.  Default is 0.05,}

\item{sigma}{Variance. Default is .20,}

\item{dt}{Timestep.  Default is 1,}

\item{T}{Total Timesteps / columns.  Default is 10,}

\item{sims}{Simulations / rows. Default is 1e4}
}
\description{
Creates a matrix simulating Brownian motion
}
\details{
The makegm function is preferred for things like pricing options.
}
\seealso{
\code{\link{makegm}},
}
\author{
Gene Leynes
}
