% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makegm.R
\name{makegm}
\alias{makegm}
\title{Make a matrix simulating Geometric Brownian motion}
\usage{
makegm(mu = 0.05, sigma = 0.2, dt = 1, T = 10, sims = 10000)
}
\arguments{
\item{mu}{Drift.  Default is 0.05,}

\item{sigma}{Variance. Default is .20,}

\item{dt}{Timestep.  Default is 1,}

\item{T}{Total Timesteps / columns.  Default is 10,}

\item{sims}{Simulations / rows. Default is 1e4}
}
\description{
Creates a matrix simulating Geometric Brownian motion
}
\details{
The makegm function is preferred over make bm for things like pricing options.
}
\author{
Gene Leynes
}
\seealso{
\code{\link{makebm}},
}

