% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot100colors.R
\name{plot100colors}
\alias{plot100colors}
\title{Generate a plot of colors to choose for plots}
\usage{
plot100colors(x)
}
\arguments{
\item{x}{Starting point for colors shown in plot}

\item{...}{Additional parameters passed to plot}
}
\description{
Plots a matrix of dots of built in colors.  The dots are labeled above
		the dot with the number and the name of the color.  You can specify
		either the name or number of the color when using in plot.
}
\details{
The first color is 1 (white), there are quite a few gradients of gray.
}
\examples{
require(geneorama)

		plot100colors(1)
		plot100colors(101)
		plot100colors(152) ## 100 shades of gray
		plot100colors(252) ## Almost 100 shades of gray
		plot100colors(352)
		plot100colors(452)
		plot100colors(552)
		plot100colors(652) ## Mostly NA (no colors after 657)
}
\author{
Gene Leynes
}
\seealso{
\code{\link[ggplot2]{ggplot}} For the underlying color function
}

