% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{seq_range}
\alias{seq_range}
\title{Sequence along range}
\usage{
seq_range(rng, ...)
}
\arguments{
\item{rng}{A vector with two elements specifying the "from" and "to" in seq}

\item{...}{further arguments passed along to seq}
}
\description{
You can feed seq_range the result of `range` i.e. a min and a max.
}
\details{
If the length of rng isn't two, seq_range will take use range(rng) to 
		calculate the range of the sequence.
}
\examples{
somedates <- as.POSIXct(c("2014-10-13", "2014-10-23", "2014-10-15"))
seq_range(range(somedates), by="day")

set.seed(1)
seq_range(range(rpois(5, 8)))
set.seed(1)
seq_range(rpois(5, 8))
set.seed(1)
seq_range(rpois(5, 8), by=.5)
set.seed(1)
seq_range(rpois(5, 8), length.out=4)


}
\author{
Gene Leynes
}
\seealso{
\code{\link[base]{seq}} Underlying function
}

