% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift a vector an arbitrary amount}
\source{
\url{http://stackoverflow.com/questions/14689424/use-a-value-from-the-previous-row-in-an-r-data-table-calculation}
}
\usage{
shift(x, offset = 1, pad = NA)
}
\arguments{
\item{x}{Vector to be shifted}

\item{pad}{Value to use for padding}

\item{shift}{Quantity to shift (forward or backward)}
}
\description{
Shift a variable by an amount, and pad with NA or your value of choice.
}
\note{
Can shift forward or backwards
}
\examples{
shift(1:10, 3)
     ## [1]  4  5  6  7  8  9 10 NA NA NA

		shift(1:10, -1)
     ## [1]  4  5  6  7  8  9 10 NA NA NA

     as.data.table(faithful)[1:5,list(eruptLengthCurrent=eruptions,
                                      eruptLengthTwoPrior=shift(eruptions,-2,0),
      								 eruptLengthThreeFuture=shift(eruptions,3))]
     # eruptLengthCurrent eruptLengthTwoPrior eruptLengthThreeFuture
     # 1:              3.600               0.000                  2.283
     # 2:              1.800               0.000                  4.533
     # 3:              3.333               3.600                     NA
     # 4:              2.283               1.800                     NA
     # 5:              4.533               3.333                     NA

     shift(1:10, 3, NaN)
     # [1]   4   5   6   7   8   9  10 NaN NaN NaN
}
\author{
Gene Leynes (originally dnlbrky \url{http://stackoverflow.com/users/1344789/dnlbrky})
}

