% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source all files within a directory}
\source{
?source
}
\usage{
sourceDir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{The path to source}

\item{trace}{Trace output}

\item{...}{Other options passed to source}
}
\description{
Run "source" on all the file in a folder.
}
\details{
This came from the example in ?source
}
\seealso{
\code{\link[base]{source}}
}
\author{
Unknown
}
