% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stirling.R
\name{stirling}
\alias{stirling}
\title{Stirling's approximation for factorials}
\usage{
stirling(x)
}
\arguments{
\item{x}{A number or vector of numbers}
}
\description{
Uses Stirling's Approximation to approximate x!
}
\details{
sqrt(2 * pi * x) * (x / exp(1)) ^ x
}
\examples{
		require(geneorama)

		factorial(10)
		stirling(10)
		
		plot(factorial(1:15), type = "o", log = "y")
		lines(stirling(1:15), type = "o", col="red")


}
\author{
Gene Leynes
}
\seealso{
\code{\link[base]{factorial}} Factorials calculated exactly.
}

