% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestGeneorama.R
\name{TestGeneorama}
\alias{TestGeneorama}
\title{TestGeneorama}
\usage{
TestGeneorama()
}
\description{
Run tests on the geneorama package to make sure the functions are
		working.
}
\details{
Depends on RUnit
}
\examples{
		\dontrun{
		library(geneorama)
		TestGeneorama()
		}

}
\author{
Gene Leynes
}

