% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Stop timing}
\usage{
toc()
}
\description{
Matlab inspired timing method.  Put tic() at the start of a process, and
		toc() at the end.
}
\details{
Creates a global variable called ticTime.
     
		System.time is preferred (see see also).
}
\seealso{
\code{\link{tic}}
	\code{\link[base]{system.time}}
}
\author{
Gene Leynes
}
