% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addbg.R
\name{addbg}
\alias{addbg}
\title{Add an opaque background to a plot when plotting}
\usage{
addbg(color = heat.colors(40, alpha = 0.15)[30], linecolor = "grey75",
  addgridlines = TRUE)
}
\arguments{
\item{color}{Background color. The default is a yellow from '
\code{heat.colors} that happens to look reasonable when 
alpha is applied.}

\item{linecolor}{Color used for gridlines.  Default is "grey75"}

\item{addgridlines}{Add gridlines? Default is TRUE}
}
\description{
Adds a transparent foreground and gridlines to plots to create the
		illusion of a background.
}
\details{
The background and gridlines can be  slightly
		more professional looking and easier to read. The color can be customized,
		the default is \code{heat.colors(40, alpha=.15)[30]}.

		This is something that's nice if you're too pressed for time to do it
		using ggplot.
		
		Uses the info in \code{par("usr")} to draw a rectangle that is the same
		size as your current plot area.  Sometimes the transparency doesn't show
		up well when copied into other applications, so use at your own risk.
}
\examples{

     require(geneorama)

     plot(1:10)
     addbg()

}
\author{
Gene Leynes
}
\seealso{
\code{\link{bgfun}},
	\code{\link{plot100colors}} to see color examples
	\code{\link[ggplot2]{ggplot}} Way better
}

