% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clipped}
\alias{clipped}
\title{Dump clipboard contents into R; especially from Excel (Windows)}
\usage{
clipped(lowercase = TRUE, removedot = TRUE)
}
\arguments{
\item{lowercase}{Converts column names to lower case}

\item{removedot}{Removes .'s from column names}
}
\description{
Useful shortcut for reading objects from your clipboard.
}
\details{
Only works in Windows. Puts everything in a data.frame, uses
		\code{stringsAsFactors=FALSE}
}
\examples{
\dontrun{

     require(geneorama)

		## Copy a range of cells in Excel (or whatever) then assign the data
		## to "clipped_example"
		clipped_example <- clipped()

		## View the result in R
		print(clipped_example)
		str(clipped_example)
		}
}
\author{
Gene Leynes
}
\seealso{
\code{\link{clipboard}},
}

