% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipper.R
\name{clipper}
\alias{clipper}
\title{Put items from R onto your cliboard (Windows)}
\usage{
clipper(
  x,
  quote = TRUE,
  sep = "\\t",
  eol = "\\n",
  na = "",
  dec = ".",
  row.names = FALSE,
  col.names = TRUE,
  qmethod = c("escape", "double")
)
}
\arguments{
\item{x}{Object to be written to clipboard.}
}
\description{
Useful shortcut to put small objects on your clipboard.  Works well
		with Excel.
}
\details{
Only works in Windows. Only works for simple objects, like 0, 1, and 2
		dimensional objects.
}
\examples{
	\dontrun{

     require(geneorama)

	## Copy a 10 by 10 matrix to your clipboard, and paste into a
	## spreadsheet or other file.
	clipper(matrix(rnorm(100),10,10))
	}

## New version based on:
## https://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r

}
\seealso{
\code{\link{clipboard}},
}
\author{
Gene Leynes
}
