% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{comma}
\alias{comma}
\title{Shortcut to add comma formatting}
\usage{
comma(x, nDigits = 2)
}
\arguments{
\item{x}{Number or vector of numbers}

\item{nDigits}{Number of trailing digits.  Default is 2}
}
\description{
Simple shortcut to make numbers into text with commas and some 
		trailing decmimal digits.  Just a shortcut to formatC.
}
\examples{
	comma(c(8987.35672, 8962.97978, 8998.06814, 8960.67199, 8997.08384))
	# [1] "8,987.36" "8,962.98" "8,998.07" "8,960.67" "8,997.08"
	comma(c(8987.35672, 8962.97978, 8998.06814, 8960.67199, 8997.08384), 0)
	# [1] "8,987" "8,963" "8,998" "8,961" "8,997"



}
\author{
Gene Leynes
}

