% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dftab}
\alias{dftab}
\title{Table as data.frame}
\usage{
dftab(x, total = TRUE, sort = TRUE)
}
\arguments{
\item{x}{Vector to tabluate}

\item{total}{Calculate total? Default is TRUE}

\item{sort}{Sort resulting data frame? Default is True}
}
\description{
Just like using \code{table} with one argument, except that \code{dftab}
		returns a sorted \code{data.frame} with a total.
}
\details{
This function is useful when you want to see a total of a table result,
		to supporess lengthy table name printing, and when you need to access
		the table result (sometimes tables can have challenges with the
		structure).  Also the NA option is hardwired to be "ifAny".
}
\examples{
require(geneorama)

		## Generate Data
		set.seed(10)
		df <- data.frame(column1 = sample(c(rep(NA, 5), rpois(25, 7))),
						 column2 = sample(letters[1:5], replace=TRUE),
						 column3 = sample(letters[1:5], replace=TRUE))
		## Tablulate each column
		dftab(df$column1)
		dftab(df$column2)
		dftab(df$column3)

		sapply(df, dftab)
}
\author{
Gene Leynes
}
\seealso{
\code{\link{table}}
}

