\name{geneorama-package}
\alias{geneorama-package}
\alias{geneorama}
\docType{package}
\title{
Gene's random collection of functions
}
\description{
Some things that I've accumulated over the years.
}
\details{
\tabular{ll}{
Package: \tab geneorama\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-07\cr
License: \tab What license is it under?\cr
}

}
\author{
Gene Leynes
Gene Leynes <gleynes@gmail.com>
}

\references{
%% TBD
}

\keyword{ package }

\seealso{
\code{\link[PBSmodelling:PBSmodelling-package]{PBSmodelling}}
}

\examples{
## TODO: Make examples? 
}
