% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inin.R
\name{inin}
\alias{inin}
\title{Table of what elements are in x and y}
\usage{
inin(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
A function that tables the degree of overlap between two vectors.
}
\details{
This will probably change.
}
\examples{
	inin(x = sample(letters[1:5], 100, TRUE), y = sample(letters[3:8], 100, TRUE))



}
\author{
Gene Leynes
}

