% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lll.R
\name{lll}
\alias{lll}
\title{lll}
\usage{
lll(sortBy = c("name", "class", "size")[1], digits = 0)
}
\arguments{
\item{sortBy}{Which column to sort output by.  One of 'name', 'class',
'size'. Default is "name"}

\item{digits}{Number of digits to show in the object size field.}
}
\description{
Nicer version of ls() with more info, and skips functions.
}
\details{
The lll function was inspired by the ll function in 'gtools', but
		is modified for different needs.  Most notibly it doesn't list functions
		so that the list is more compact, and it's easier to see what's
		taking up space in your environment.
}
\examples{
require(geneorama)
	lll()
}
\author{
Gene Leynes
}
\seealso{
\code{\link{ll}} (in gdata)
}

