% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{loader (DEPRICATED)}
\alias{loader}
\alias{loader (DEPRICATED)}
\title{Loads a single object into memory}
\usage{
loader(x, xPath = ".")
}
\arguments{
\item{x}{filename}

\item{xPath}{filepath}
}
\description{
This was before I knew about readRDS.
}
\details{
Use readRDS instead
}
\author{
Gene Leynes
}
\seealso{
\code{\link{saver}},
	\code{\link[base]{readRDS.}}
}

