% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadinstall_libraries.R
\name{loadinstall_libraries}
\alias{loadinstall_libraries}
\title{Load / Install CRAN Packages}
\usage{
loadinstall_libraries(required_packages = c(),
  myrepos = "http://cran.rstudio.com/", install_only = FALSE)
}
\arguments{
\item{required_packages}{Character vector of pacakges to load / install
The default is "", in which case it doesn't try 
to install anything.}

\item{myrepos}{Repository to use. Default is 
"http://cran.rstudio.com/"}

\item{install_only}{Boolean to indicate if the packages are only 
to be installed, instead of installed an loaded.}
}
\description{
Call this function to load a vector of required packages from
		CRAN. It will try to load the packages, and intall them if they are
		missing.
}
\details{
You may wish to update the default repo: "http://cran.rstudio.com/"
	##
	## 2014-03-31   SJR   Extracted this function from 00_Initialize.R
	## 2014-05-05   GWL   Adding function to geneorama package
	##
}
\examples{
		## Warning: this will install these packages if you don't already have
		## them
		\dontrun{
	require(geneorama)
	loadinstall_libraries(c("ggplot2", "data.table"))
		}
 require(geneorama)
lll()

}
\author{
Gene Leynes
}

