% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDebugMessage.R
\name{printDebugMessage}
\alias{printDebugMessage}
\title{Embed within functions to optionally print debug mesagges at run time}
\usage{
printDebugMessage(DebugMode, msg = NULL, indent = 0)
}
\arguments{
\item{DebugMode}{Are you in debug mode now?  Default is FALSE}

\item{msg}{Message to print.  Default is NULL}

\item{indent}{How far should msg be indented?  Default is 0, but 
it's also modified internally}
}
\description{
Prints messages at runtime, optionally, at various levels to help debug.
}
\details{
some details are... comming soon.
}
\author{
Gene Leynes
}

