% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saver (DEPRICATED)}
\alias{saver}
\alias{saver (DEPRICATED)}
\title{Saves a single object to a file}
\usage{
saver(x, override = FALSE)
}
\arguments{
\item{x}{filename}

\item{override}{overwrite file if it exists?  Default is FALSE.}
}
\description{
This was before I knew about saveRDS.
}
\details{
Use saveRDS instead
}
\author{
Gene Leynes
}
\seealso{
\code{\link{loader}},
	\code{\link[base]{saveRDS.}}
}

