% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf.R
\name{wtf}
\alias{wtf}
\title{"Write Temp Function" will write a matrix / data.table / data.frame
        to a temp csv file and attempt to open it with the registered 
        application (hopefully in Excel)}
\usage{
wtf(x)
}
\description{
"Write Temp Function" will write a matrix / data.table / data.frame
		to a temp csv file and attempt to open it with the registered 
		application (hopefully in Excel)
}
\details{
Has some logic to avoid name clashes when opening two objects of the
     same name (e.g. say you open temp, change it, then open temp again)
}
\author{
Gene Leynes
}

